# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ETask API Documentation",
    "description": "基于 AST 与反射契约混合自适应生成的 OpenAPI 3.0 规范，支持导入 Apifox / Postman 进行全量接口联调。",
    "version": "1.0.0"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/codebook/children": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "代码资源子节点",
        "description": "权限标识: children",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildrenReq"
              },
              "example": {
                "parent_id": 1,
                "project_id": 1,
                "scope": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/codebook/create": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "创建模板",
        "description": "权限标识: add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReq"
              },
              "example": {
                "key": "string",
                "secret": true,
                "value": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/codebook/delete/{id}": {
      "delete": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "删除脚本模板",
        "description": "权限标识: delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/codebook/detail/{id}": {
      "get": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "模板详情",
        "description": "权限标识: get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/codebook/file/{id}/download": {
      "get": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "下载项目文件",
        "description": "权限标识: download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
    "/api/codebook/rename": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "重命名模板",
        "description": "权限标识: rename",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameReq"
              },
              "example": {
                "id": 1,
                "name": "示例名称"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/codebook/sort": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "模板排序",
        "description": "权限标识: sort",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SortReq"
              },
              "example": {
                "id": 1,
                "target_parent_id": 1,
                "target_position": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/codebook/tree/{project_id}": {
      "get": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "代码资源树",
        "description": "权限标识: view_tree",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/codebook/update": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "更新模板",
        "description": "权限标识: edit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReq"
              },
              "example": {
                "id": 1,
                "key": "string",
                "secret": true,
                "value": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/codebook/workspace/file": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "读取制品文件",
        "description": "权限标识: view_workspace_tree",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceFileReq"
              },
              "example": {
                "artifact_path": "string",
                "digest": "string",
                "project_id": 1,
                "release_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
    "/project/archive/{id}": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "归档项目",
        "description": "权限标识: delete_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/project/create": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "创建项目",
        "description": "权限标识: add_project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectReq"
              },
              "example": {
                "artifact_enabled": true,
                "artifact_namespace": "示例名称",
                "desc": "相关业务描述信息",
                "name": "示例名称"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/project/delete-impact/{id}": {
      "get": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "项目删除影响",
        "description": "权限标识: project_delete_impact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/project/delete/{id}": {
      "delete": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "删除项目",
        "description": "权限标识: purge_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDeleteReq"
              },
              "example": {
                "project_name": "示例名称"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/project/detail/{id}": {
      "get": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "项目详情",
        "description": "权限标识: get_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/project/list": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "项目列表",
        "description": "权限标识: view_project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListProjectsReq"
              },
              "example": {
                "keyword": "search_key",
                "scope": "string",
                "status": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/project/references": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "可引用项目列表",
        "description": "权限标识: reference_projects",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListReferenceProjectsReq"
              },
              "example": {
                "exclude_project_id": 1,
                "keyword": "search_key"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/project/restore/{id}": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "恢复项目",
        "description": "权限标识: restore_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/project/update": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "更新项目",
        "description": "权限标识: edit_project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectReq"
              },
              "example": {
                "artifact_enabled": true,
                "artifact_namespace": "示例名称",
                "desc": "相关业务描述信息",
                "id": 1,
                "name": "示例名称",
                "sort_no": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/version/create": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "创建版本",
        "description": "权限标识: add_version",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVersionReq"
              },
              "example": {
                "code": "admin",
                "message": "string",
                "node_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/version/detail/{id}": {
      "get": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "版本详情",
        "description": "权限标识: get_version",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/version/list": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "版本列表",
        "description": "权限标识: view_version",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListVersionsReq"
              },
              "example": {
                "node_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/version/use": {
      "post": {
        "tags": [
          "脚本引擎/脚本模板"
        ],
        "summary": "使用版本",
        "description": "权限标识: use_version",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UseVersionReq"
              },
              "example": {
                "node_id": 1,
                "version_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    }
  },
...
  "components": {
    "schemas": {
...
      "ChildrenReq": {
        "type": "object",
        "properties": {
          "parent_id": {
            "type": "integer",
            "example": 1
          },
          "project_id": {
            "type": "integer",
            "example": 1
          },
          "scope": {
            "type": "string",
            "example": "string"
          }
        },
        "example": {
          "parent_id": 1,
          "project_id": 1,
          "scope": "string"
        }
      },
...
      "CreateReq": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "string"
          },
          "secret": {
            "type": "boolean",
            "example": true
          },
          "value": {
            "type": "string",
            "example": "string"
          }
        },
        "example": {
          "key": "string",
          "secret": true,
          "value": "string"
        }
      },
...
      "ProjectDeleteReq": {
        "type": "object",
        "properties": {
          "project_name": {
            "type": "string",
            "example": "示例名称"
          }
        },
        "example": {
          "project_name": "示例名称"
        }
      },
...
      "RenameReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "示例名称"
          }
        },
        "example": {
          "id": 1,
          "name": "示例名称"
        }
      },
...
      "SortReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "target_parent_id": {
            "type": "integer",
            "example": 1
          },
          "target_position": {
            "type": "integer",
            "example": 0
          }
        },
        "example": {
          "id": 1,
          "target_parent_id": 1,
          "target_position": 0
        }
      },
...
      "UpdateReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "key": {
            "type": "string",
            "example": "string"
          },
          "secret": {
            "type": "boolean",
            "example": true
          },
          "value": {
            "type": "string",
            "example": "string"
          }
        },
        "example": {
          "id": 1,
          "key": "string",
          "secret": true,
          "value": "string"
        }
      },
...
      "WorkspaceFileReq": {
        "type": "object",
        "properties": {
          "artifact_path": {
            "type": "string",
            "example": "string"
          },
          "digest": {
            "type": "string",
            "example": "string"
          },
          "project_id": {
            "type": "integer",
            "example": 1
          },
          "release_id": {
            "type": "integer",
            "example": 1
          }
        },
        "example": {
          "artifact_path": "string",
          "digest": "string",
          "project_id": 1,
          "release_id": 1
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "请输入登录后获取的 JWT Bearer Token 进行接口调用鉴权"
      }
    }
  }
}